refactor: share service trace signal normalization and aggregation
Extract shared helpers for trace status label normalization and slow operation aggregation, then reuse them in both ServiceDashboardPage insight context and ServiceTraceBreakdownPanels so panel rendering and insight signals stay aligned.

Made-with: Cursor
